refactor: execAcceptOfferTx and execWithdrawOfferTx
