+++ composables/transaction/transactionOfferWithdraw.ts
@@ -1,23 +1,3 @@
-import { warningMessage } from '@/utils/notification'
-import { tokenIdToRoute } from '@/components/unique/utils'
+import { transactionOfferFactory } from './utils'
 
-import type { ActionWithdrawOffer } from './types'
-
-export function execWithdrawOfferTx(
-  params: ActionWithdrawOffer,
-  api,
-  executeTransaction
-) {
-  try {
-    const { id, item } = tokenIdToRoute(params.nftId)
-    const args = [id, item, params.maker]
-    const cb = api.tx.marketplace.withdrawOffer
-
-    executeTransaction({
-      cb,
-      arg: args,
-    })
-  } catch (error) {
-    warningMessage(error)
-  }
-}
+export const execWithdrawOfferTx = transactionOfferFactory('withdrawOffer')
